diplom_store/online_store/pyproject.toml
[tool.poetry]
name = "diplom-store"
version = "0.1.0"
description = ""
authors = ["ramazan_aisaev <aisaev-68@yandex.ru>"]
readme = "README.md"
packages = [{include = "diplom_store"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
python-decouple = "^3.8"
djangorestframework = "^3.14.0"
pillow = "^9.5.0"
drf-yasg = "^1.21.5"
docutils = "^0.19"
scrapy = "^2.9.0"
selenium = "^4.9.1"
scrapy-user-agents = "^0.1.1"
django-filter = "^23.2"
django-cors-headers = "^4.0.0"
mimesis = "^10.1.0"
coreapi = "^2.3.3"
django-modeltranslation = "^0.18.10"
xlrd = "^2.0.1"
xlwt = "^1.3.0"
psycopg2-binary = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
